[cryptography] Improve comments (#275)
