+++ cryptography/Cargo.toml
@@ -33,14 +33,14 @@ prost-build = { workspace = true }
 portable = ["blst/portable"]
 
 [[bench]]
-name="idkg_recovery"
+name="dkg_recovery"
 harness = false
-path = "src/bls12381/benches/idkg_recovery.rs"
+path = "src/bls12381/benches/dkg_recovery.rs"
 
 [[bench]]
-name="idkg_reshare_recovery"
+name="dkg_reshare_recovery"
 harness = false
-path = "src/bls12381/benches/idkg_reshare_recovery.rs"
+path = "src/bls12381/benches/dkg_reshare_recovery.rs"
 
 [[bench]]
 name="signature_generation"
