+++ cryptography/src/bls12381/benches/partial_signature_aggregation.rs
@@ -1,4 +1,4 @@
-use commonware_cryptography::bls12381::{idkg, primitives};
+use commonware_cryptography::bls12381::{dkg, primitives};
 use commonware_utils::quorum;
 use criterion::{criterion_group, criterion_main, BatchSize, Criterion};
 use std::hint::black_box;
@@ -11,7 +11,7 @@ fn benchmark_partial_signature_aggregation(c: &mut Criterion) {
         c.bench_function(&format!("n={} t={}", n, t), |b| {
             b.iter_batched(
                 || {
-                    let (_, shares) = idkg::ops::generate_shares(None, n, t);
+                    let (_, shares) = dkg::ops::generate_shares(None, n, t);
                     shares
                         .iter()
                         .map(|s| primitives::ops::partial_sign(s, namespace, msg))
