+++ docs/blogs/commonware-cryptography.html
@@ -19,7 +19,7 @@
     <meta property="og:description" content="When I chat with a developer about launching their own onchain application (as a specialized blockchain), the same questions always come up." />
     <meta property="article:author" content="https://x.com/_patrickogrady" />
     <meta property="article:published_time" content="2024-08-28T00:00:00Z" />
-    <meta property="article:modified_time" content="2024-12-13T00:00:00Z" />
+    <meta property="article:modified_time" content="2024-12-14T00:00:00Z" />
 
     <meta name="twitter:card" content="summary_large_image" />
     <meta property="twitter:domain" content="commonware.xyz" />
@@ -92,8 +92,8 @@
             </li>
         </ul>
         <p>Today, we're excited to share how @commonwarexyz is answering these questions with <a href="https://docs.rs/commonware-cryptography">commonware-cryptography</a> (now in ALPHA).</p>
-        <p>commonware-cryptography is an open (Apache-2 and MIT) implementation of BLS12-381 Distributed Key Generation (DKG), Resharing, and Threshold Signatures in Rust. Unlike most implementations, commonware-cryptography does not employ a "board" for share distribution (that requires dealers to broadcast encrypted shares over a public channel). Rather, dealers distribute shares over encrypted connections directly with each recipient and recipients broadcast an acknowledgement when they receive a correct share (no share material included).</p>
-        <p>This interactive DKG/Resharing construction supports contributor disqualification (missing commitment, invalid commitment, invalid share distribution, ignoring reveal requests, etc.) and can maintain a threshold of f + 1 (of 3f + 1 total contributors) under the partially synchronous network model (where messages may be arbitrarily delayed between any 2 participants) across reshares (which can be run to introduce/remove contributors and/or proactively on a regular interval to bound the age of shares). You can read more about this construction <a href="https://docs.rs/commonware-cryptography/latest/commonware_cryptography/bls12381/dkg/index.html">here</a>.</p>
+        <p>commonware-cryptography is an open (Apache-2 and MIT) implementation of BLS12-381 Distributed Key Generation (DKG), Resharing, and Threshold Signatures in Rust. Unlike most implementations, our construction does not employ a "board" for share distribution (that requires dealers to broadcast encrypted shares over a public channel). Rather, dealers distribute shares over encrypted connections directly with each recipient and recipients broadcast an acknowledgement when they receive a correct share (no share material included).</p>
+        <p>This interactive DKG/Resharing construction supports contributor disqualification (invalid commitment, invalid share generation, etc.) and can maintain a threshold of f + 1 (of 3f + 1 total contributors) under the partially synchronous network model (where messages may be arbitrarily delayed between any 2 participants) across reshares. Reshares can be run to introduce/remove contributors and/or proactively on a regular interval to bound the age of any shares. You can read more about this construction <a href="https://docs.rs/commonware-cryptography/latest/commonware_cryptography/bls12381/dkg/index.html">here</a>.</p>
         <p>So, how does this answer any of a developer's questions?</p>
         <ul>
             <li>
@@ -106,10 +106,7 @@
                 <p><b>Views <i>(Access)</i>:</b> "Links" can be extended beyond blockchain execution environments to wallets and explorers to provide trust-minimized access to onchain activity. Like a blockchain, all a user requires to verify that a blockchain included/emitted some data is its static public key (can be hardcoded) and a Threshold Signature over a block (or whatever artifact a developer wants to emit) referencing said data.</p>
             </li>
         </ul>
-        <p>To showcase how "Seeds" will work, we implemented a <a href="https://docs.rs/commonware-vrf">Proof-of-Concept Verifiable Random Function (VRF) CLI</a> with a trusted Arbiter (that will be removed after the release of commonware-consensus) and untrusted Contributors:</p>
-        <div class="video-container">
-            <iframe src="https://www.youtube.com/embed/nGIJk4w63F0?si=jy2QGEvsC0WHJUwi" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
-        </div>
+        <p>To showcase how "Seeds" will work, we implemented a <a href="https://docs.rs/commonware-vrf">Proof-of-Concept Verifiable Random Function (VRF) CLI</a> with a trusted Arbiter and untrusted Contributors.</p>
         <p>Contributors to this VRF connect to each other over commonware-p2p (using ED25519 identities), perform an initial DKG (to generate a static public key), and then perform a proactive refresh every 10 seconds. After a successful DKG and/or Reshare, contributors generate partial signatures over the round number and gossip them to others in the group (again using commonware-p2p). These partial signatures, when aggregated, form a bias-resistant source of randomness that was not knowable to any contributor prior to signing.</p>
         <p>To demonstrate how malicious contributors are tolerated, the CLI also lets you behave as a "rogue" dealer that generates invalid shares or a "lazy" dealer that doesn't distribute shares to other contributors.</p>
         <p>Next up, commonware-consensus!</p>
